requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
einops==0.8.0
ftfy==6.2.3
google-auth==2.37.0
google-auth-oauthlib==1.0.0
grpcio==1.68.1
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Markdown==3.7
MarkupSafe==2.1.5
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.2
pathfinding==1.0.16
pillow==10.4.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.10.1
six==1.17.0
smplx==0.1.28
tensorboard==2.14.0
tensorboard-data-server==0.7.2
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
vit-pytorch==1.6.5
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.20.2
